requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Cython==3.0.12
distro==1.9.0
docutils==0.21.2
filetype==1.2.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
Kivy==2.3.1
Kivy-Garden==0.1.5
macholib==1.16.3
MarkupSafe==3.0.2
openai==1.63.0
osascript==2020.12.3
packaging==24.2
pillow==11.1.0
psutil==6.1.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
python-dotenv==1.1.0
requests==2.32.3
runcmd==2020.12.3
setuptools==75.8.0
sniffio==1.3.1
tqdm==4.67.1
turtler==0.0.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
